Deleted Tube class; merged Plate into Labware class
Manually edited some old migrations, so they have to be run from
scratch.
