--- app/models/labwares/tube.rb
@@ -1,3 +1,0 @@
-class Tube < Labware
-  belongs_to :labware_type
-end
